# [SPARK-9141][SQL] Remove project collapsing from DataFrame API