… sql/catalyst/src/main/scala/org/apache/spark/sql/catalyst/plans/logical/LogicalPlan.scala
… sql/catalyst/src/main/scala/org/apache/spark/sql/catalyst/rules/RuleExecutor.scala
… sql/catalyst/src/main/scala/org/apache/spark/sql/catalyst/trees/TreeNode.scala
… sql/catalyst/src/test/scala/org/apache/spark/sql/catalyst/plans/LogicalPlanSuite.scala
… sql/core/src/main/scala/org/apache/spark/sql/DataFrame.scala
… sql/core/src/main/scala/org/apache/spark/sql/execution/SparkPlan.scala
… sql/core/src/main/scala/org/apache/spark/sql/execution/pythonUDFs.scala
… sql/core/src/test/scala/org/apache/spark/sql/CachedTableSuite.scala
… sql/core/src/test/scala/org/apache/spark/sql/DataFrameSuite.scala
… sql/hive/compatibility/src/test/scala/org/apache/spark/sql/hive/execution/HiveCompatibilitySuite.scala
… sql/hive/src/main/scala/org/apache/spark/sql/hive/hiveUDFs.scala
… sql/hive/src/test/scala/org/apache/spark/sql/hive/execution/HiveUDFSuite.scala
… sql/hive/src/test/scala/org/apache/spark/sql/hive/execution/SQLQuerySuite.scala