@cloud-fan 2015-08-04
It's kind of annoying to check `analyzed` at every rule in `Analyzer` and all related places, how about we improve the semantic of plan equality? i.e. enhance `LogicalPlan.sameResult`.
cc @rxin