@marmbrus 2015-08-04
It would be good to improve `sameResult`, but I think that is much to scary of a change to merge in after feature freeze.  Query subsumption is hard, and while what we do today is limited, it is simple and easy to reason about.  Past attempts to improve this function have had subtle correctness bugs.

Regarding checking in a lot of places, we already have to check for `child.resolved` in many places already.  Depending on time, I'll look at migrating this logic into a special version of transform, where we can be even more efficient.