@yhuai 2015-08-05
Looks good to me. Left a few minor comments related to comments and style.