@yhuai 2015-08-05
I am merging it to master and branch 1.5. Will send out a follow-up PR to fix comments and style.